+++ karma.cross-browser.config.js
@@ -1,3 +1,5 @@
+var ExtractTextPlugin = require('extract-text-webpack-plugin');
+
 module.exports = function(config) {
 
   /**
@@ -59,13 +61,17 @@ module.exports = function(config) {
      * The configuration for the karma-webpack plugin.
      */
     webpack: {
+      plugins: [
+        new ExtractTextPlugin('style.css', { allChunks: true })
+      ],
       module: {
         loaders: [
-          { test: /\.jsx?$/, exclude: /node_modules/, loader: "babel-loader?stage=0"}
+          { test: /\.jsx?$/, exclude: /node_modules/, loader: "babel-loader?stage=0" },
+          { test: /\.css$/, loader: ExtractTextPlugin.extract('style-loader', 'css-loader?modules&importLoaders=1&localIdentName=[name]__[local]___[hash:base64:5]!postcss-loader') }
         ]
       },
       resolve: {
-        extensions: ['', '.js', '.jsx']
+        extensions: ['', '.js', '.jsx', '.css']
       }
     },