+++ src/pages/landing/page.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { getData } from "../../common/request";
+import styles from "./style.css";
 
 
 export default class LandingPage extends React.Component {
@@ -11,10 +12,16 @@ export default class LandingPage extends React.Component {
     let { title } = this.props.data.landing;
 
     return (
-      <div id="landing-page">
-        <h1>{title}</h1>
+      <div className={styles.content}>
+        <h1 className={styles.heading}>{title}</h1>
+        <p className={styles.lead}>Create an account to get started!</p>
+        <button className={styles.signUpButton} onClick={this.signUp}>Sign up</button>
       </div>
     );
+  }
+
+  signUp = (event) => {
+    alert("Sign Up!");
   }
 
   static fetchData = function(params) {