+++ src/routers/logged_out.jsx
@@ -8,9 +8,7 @@ export default class LoggedOutRouter extends React.Component {
   render() {
     return (
       <div id="container">
-        <div id="main">
-          <RouteHandler {...this.props} />
-        </div>
+        <RouteHandler {...this.props} />
       </div>
     );
   }