+++ webpack.production.config.js
@@ -1,4 +1,5 @@
 var webpack = require('webpack');
+var ExtractTextPlugin = require('extract-text-webpack-plugin');
 
 /**
  * This is the Webpack configuration file for production.
@@ -11,13 +12,23 @@ module.exports = {
     filename: "app.js"
   },
 
+  plugins: [
+    new ExtractTextPlugin('style.css', { allChunks: true })
+  ],
+
   module: {
     loaders: [
-      { test: /\.jsx?$/, exclude: /node_modules/, loader: "babel-loader?stage=0" }
+      { test: /\.jsx?$/, exclude: /node_modules/, loader: "babel-loader?stage=0" },
+      { test: /\.css$/, loader: ExtractTextPlugin.extract('style-loader', 'css-loader?modules&importLoaders=1&localIdentName=[name]__[local]___[hash:base64:5]!postcss-loader') }
     ]
   },
 
   resolve: {
-    extensions: ['', '.js', '.jsx']
-  }
+    extensions: ['', '.js', '.jsx', '.css']
+  },
+
+  postcss: [
+    require('autoprefixer'),
+    require('postcss-nested')
+  ]
 }